Update JJs_HashMonitor.ps1
Ocassionally a card will begin hashing at dropped hash rate or OverdriveNTool will fail to set a card. In the later case only the CPU will be hashing, because a card failed to disable/enable and this script in that case won't initialize the cards. I put in a user defined variable, $global:minHash, that I set to 100 H/s lower than my normal 15min hash rate. This will catch a drop in normal performance and restart STAK. This can get stuck in a loop if one or more cards always fail to disable/enable. I'm going to try to add in a check for this and if it's happening, a restart of the computer usually fixes it, as long as you have script scheduled to run at Windows start.